+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - inputs which would be overwritten by a compendium link are disabled
+- equipped apparel now applies actor targeted rule elements to the actor
 
 ## [0.14.3] - 2022-03-30